removed a bunch of the component/common files as they arent in use as widely as planned.
