+++ packages/augur-ui/src/modules/modal/components/common/modal-actions.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ModalButton from "modules/modal/components/common/modal-button";
+import classNames from "classnames";
 
 import Styles from "modules/modal/components/common/common.styles.less";
 
@@ -14,6 +14,43 @@ interface ButtonObject {
 interface ModalActionsProps {
   buttons: ButtonObject[];
 }
+
+interface ModalButtonProps {
+  label: string;
+  action: Function;
+  type?: string;
+  continueDefault?: boolean;
+  isDisabled?: boolean;
+}
+
+const ModalButton = ({ type = "purple", label, action, isDisabled = false, continueDefault = false }: ModalButtonProps) => {
+  let styleType = `${Styles["ActionButtons__button-purple"]}`;
+  switch (type) {
+    case "gray":
+    case "grey":
+      styleType = `${Styles["ActionButtons__button-gray"]}`;
+      break;
+    case "offWhite":
+      styleType = `${Styles["ActionButtons__button-offWhite"]}`;
+      break;
+    default:
+      break;
+  }
+  return (
+    <button
+      className={classNames(Styles.ActionButtons__button, {
+        [`${styleType}`]: true,
+      })}
+      disabled={!!isDisabled}
+      onClick={(e) => {
+        if (!continueDefault) e.preventDefault();
+        action(e);
+      }}
+    >
+      {label}
+    </button>
+  );
+};
 
 const ModalActions: React.FC<ModalActionsProps> = ({ buttons }) => (
   <div className={Styles.ActionButtons}>
