--- packages/augur-ui/src/modules/modal/components/common/modal-description.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import ReactHtmlParser from "react-html-parser";
-
-import Styles from "modules/modal/components/common/common.styles.less";
-
-interface ModalDescriptionProps {
-  text: string;
-}
-
-const ModalDescription = ({ text }: ModalDescriptionProps) => (
-  <p className={Styles.Description}>{ReactHtmlParser(text)}</p>
-);
-
-export default ModalDescription;
